helper tests in separate file
